Add PHPUnit tests

+++ lib/Controller/SettingsController.php
@@ -80,7 +80,8 @@ class SettingsController extends Controller {
      *
      */
     public function addClient() {
-        if (filter_var($_POST['redirect_uri'], FILTER_VALIDATE_URL) === false) {
+		if (!isset($_POST['redirect_uri']) || !isset($_POST['name']) ||
+			filter_var($_POST['redirect_uri'], FILTER_VALIDATE_URL) === false) {
             return new RedirectResponse('../../settings/admin#oauth-2.0');
         }
 
@@ -98,7 +99,7 @@ class SettingsController extends Controller {
     /**
      * Deletes a client.
 	 *
-	 * @param string $id The client identifier.
+	 * @param int $id The client identifier.
      *
      * @return RedirectResponse Redirection to the settings page.
      *
@@ -106,10 +107,16 @@ class SettingsController extends Controller {
      *
      */
     public function deleteClient($id) {
+		if (!is_int($id)) {
+			return new RedirectResponse('../../../../settings/admin#oauth-2.0');
+		}
+
         $client = $this->clientMapper->find($id);
         $this->clientMapper->delete($client);
 
-        // TODO: Delete all Authorization Codes, Access Tokens and Refresh Tokens
+        $this->authorizationCodeMapper->deleteByClient($id);
+        $this->accessTokenMapper->deleteByClient($id);
+        $this->refreshTokenMapper->deleteByClient($id);
 
         return new RedirectResponse('../../../../settings/admin#oauth-2.0');
     }
@@ -117,7 +124,7 @@ class SettingsController extends Controller {
 	/**
 	 * Revokes the authorization for a client.
 	 *
-	 * @param string $id The client identifier.
+	 * @param int $id The client identifier.
 	 * @param string $user_id The ID of the user logged in.
 	 *
 	 * @return RedirectResponse Redirection to the settings page.
@@ -126,6 +133,10 @@ class SettingsController extends Controller {
 	 *
 	 */
 	public function revokeAuthorization($id, $user_id) {
+		if (!is_int($id) || !is_string($user_id)) {
+			return new RedirectResponse('../../../../settings/personal#oauth-2.0');
+		}
+
 		$this->authorizationCodeMapper->deleteByClientUser($id, $user_id);
 		$this->accessTokenMapper->deleteByClientUser($id, $user_id);
 		$this->refreshTokenMapper->deleteByClientUser($id, $user_id);
