+++ lib/Db/AccessTokenMapper.php
@@ -25,6 +25,7 @@
 namespace OCA\OAuth2\Db;
 
 use InvalidArgumentException;
+use OCA\OAuth2\Controller\SettingsController;
 use OCP\AppFramework\Db\Entity;
 use OCP\IDb;
 use OCP\AppFramework\Db\Mapper;
@@ -52,7 +53,27 @@ class AccessTokenMapper extends Mapper {
 		}
 
 		$sql = 'SELECT * FROM `' . $this->tableName . '` WHERE `id` = ?';
-		return $this->findEntity($sql, array($id), null, null);
+		return $this->findEntity($sql, [$id], null, null);
+	}
+
+	/**
+	 * Selects an access token by its token.
+	 *
+	 * @param string $token The access token.
+	 *
+	 * @return Entity The access token entity.
+	 *
+	 * @throws \OCP\AppFramework\Db\DoesNotExistException if not found.
+	 * @throws \OCP\AppFramework\Db\MultipleObjectsReturnedException if more
+	 * than one result.
+	 */
+	public function findByToken($token) {
+		if (!is_string($token)) {
+			throw new InvalidArgumentException('Argument token must be a string');
+		}
+
+		$sql = 'SELECT * FROM `'. $this->tableName . '` WHERE `token` = ?';
+		return $this->findEntity($sql, [$token], null, null);
 	}
 
 	/**
@@ -79,10 +100,35 @@ class AccessTokenMapper extends Mapper {
 			throw new InvalidArgumentException('Argument client_id must be an int and user_id must be a string');
 		}
 
-		$sql = 'DELETE FROM `' . $this->tableName . '` '
-			. 'WHERE client_id = ? AND user_id = ?';
-		$stmt = $this->execute($sql, array($clientId, $userId), null, null);
+		$sql = 'DELETE FROM `' . $this->tableName . '` ' . 'WHERE `client_id` = ? AND `user_id` = ?';
+		$stmt = $this->execute($sql, [$clientId, $userId], null, null);
 		$stmt->closeCursor();
 	}
 
+	/**
+	 * Deletes all entities from the table
+	 */
+	public function deleteAll(){
+		$sql = 'DELETE FROM `' . $this->tableName . '`';
+		$stmt = $this->execute($sql, []);
+		$stmt->closeCursor();
+	}
+
+    /**
+     * Deletes all access tokens for a given client_id.
+     * Used for client deletion by the administrator in the
+     * admin settings.
+     *
+     * @param int $clientId The client ID
+     * @see SettingsController::deleteClient()
+     */
+	public function deleteByClient($clientId) {
+        if (!is_int($clientId)) {
+            throw new InvalidArgumentException('Argument client_id must be an int');
+        }
+
+        $sql = 'DELETE FROM `' . $this->tableName . '` ' . ' WHERE `client_id` = ?';
+        $stmt = $this->execute($sql, [$clientId], null, null);
+        $stmt->closeCursor();
+    }
 }
