+++ lib/Db/AuthorizationCodeMapper.php
@@ -52,7 +52,7 @@ class AuthorizationCodeMapper extends Mapper {
 		}
 
         $sql = 'SELECT * FROM `'. $this->tableName . '` WHERE `id` = ?';
-        return $this->findEntity($sql, array($id), null, null);
+        return $this->findEntity($sql, [$id], null, null);
     }
 
     /**
@@ -72,7 +72,7 @@ class AuthorizationCodeMapper extends Mapper {
 		}
 
         $sql = 'SELECT * FROM `'. $this->tableName . '` WHERE `code` = ?';
-        return $this->findEntity($sql, array($code), null, null);
+        return $this->findEntity($sql, [$code], null, null);
     }
 
     /**
@@ -99,10 +99,36 @@ class AuthorizationCodeMapper extends Mapper {
 			throw new InvalidArgumentException('Argument client_id must be an int and user_id must be a string');
 		}
 
-		$sql = 'DELETE FROM `' . $this->tableName . '` '
-			. 'WHERE client_id = ? AND user_id = ?';
-		$stmt = $this->execute($sql, array($clientId, $userId), null, null);
+		$sql = 'DELETE FROM `' . $this->tableName . '` ' . 'WHERE `client_id` = ? AND `user_id` = ?';
+		$stmt = $this->execute($sql, [$clientId, $userId], null, null);
 		$stmt->closeCursor();
 	}
+
+	/**
+	 * Deletes all entities from the table
+	 */
+	public function deleteAll() {
+		$sql = 'DELETE FROM `' . $this->tableName . '`';
+		$stmt = $this->execute($sql, []);
+		$stmt->closeCursor();
+	}
+
+    /**
+     * Deletes all authorization codes for a given client ID.
+     * Used for client deletion by the administrator in the
+     * admin settings.
+     *
+     * @param int $clientId The client ID
+     * @see SettingsController::deleteClient()
+     */
+    public function deleteByClient($clientId) {
+        if (!is_int($clientId)) {
+            throw new InvalidArgumentException('Argument client_id must be an int');
+        }
+
+        $sql = 'DELETE FROM `' . $this->tableName . '` ' . 'WHERE `client_id` = ?';
+        $stmt = $this->execute($sql, [$clientId], null, null);
+        $stmt->closeCursor();
+    }
 
 }
